+++ control-plane/Cargo.toml
@@ -26,6 +26,8 @@ aws-config = "0.55.0"
 aws-types = "0.55.0"
 aws-sdk-sns = "0.25.0"
 semver = "1.0.17"
+cadence = "0.29.0"
+cadence-macros = "0.29.0"
 
 [dev-dependencies]
 tokio-test = "0.4.2"
