+++ control-plane/src/lib.rs
@@ -11,4 +11,5 @@ pub mod enclave_connection;
 pub mod error;
 pub mod health;
 pub mod internal_dns;
+pub mod stats_client;
 pub mod stats_proxy;
