+++ shared/Cargo.toml
@@ -24,6 +24,8 @@ rand = "^0.8"
 tls-parser = { version = "0.11.0", optional = true }
 ppp = { version = "2.2.0" }
 tokio-rustls = { version = "0.23.4", features = ["dangerous_configuration"] }
+sys-info = "0.9.1"
+cadence = "0.29.0"
 
 [dev-dependencies]
 tokio-test = "0.4.2"
